requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.9.2

PyMuPDF==1.24.14
python-docx==1.1.2

sentence-transformers==3.2.1
faiss-cpu==1.9.0.post1
numpy==1.26.4

openai==1.57.4
transformers==4.46.3
torch

